[ja_jp] Apply changes from the original to v2.5
This huge patch applies changes in the original repository (fabric) from
v2.2 to v2.5 to the Japanese documentation in the release-2.5 branch.
This patch adds, modifies and deletes the binary and text files from the
original repository except the text files translated in the release-2.2,
because the files not applied in this patch should be manually applied
with translation to Japanese.

Signed-off-by: Taku Shimosawa <taku.shimosawa.uf@hitachi.com>